# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Goggins Habit Tracker API
  description: Backend API for the Goggins Habit Tracker application, providing authentication, data persistence, and AI proxies.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        api_key:
          type: string
          description: "Encrypted Gemini API Key provided by the user"
    LoginRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
        api_key:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'

    # --- Core Data Schemas (from types.ts) ---
...
    Goal:
      type: object
      required: [id, description, targetDate]
      properties:
        id:
          type: string
        description:
          type: string
        targetDate:
          type: string
          format: date
        label:
          type: string
        completed:
          type: boolean
        completionDate:
          type: string
        completionProof:
          type: string
        completionFeedback:
          type: string
        system:
          $ref: '#/components/schemas/AtomicHabitsSuggestions'
        contract:
          $ref: '#/components/schemas/GoalContract'
    WeeklyGoal:
      type: object
      required: [id, description, targetDate]
      properties:
        id: 
          type: string
        description: 
          type: string
        targetDate: 
          type: string
        alignedGoalId:
          type: string
        goalAlignment:
          type: number
        label:
          type: string
        evaluation:
          $ref: '#/components/schemas/WeeklyGoalEvaluation'
        contract:
          $ref: '#/components/schemas/GoalContract'
    GoalContract:
      type: object
      properties:
        primaryObjective: {type: string}
        contractStatement: {type: string}
        rewardPayout: {type: number}
        kpis:
          type: array
          items:
            type: object
            properties:
              description: {type: string}
              type: {type: string, enum: ['Internal Metric', 'External Metric']}
              target: {type: string}
        preStateAnswers:
          type: array
          items:
            type: object
            properties:
              question: {type: string}
              answer: {type: string}
        fiveWhys:
          type: array
          items: {type: string}
...
    Reward:
      type: object
      properties:
        id: {type: string}
        name: {type: string}
        cost: {type: number}
    PurchasedReward:
      type: object
      properties:
         id: {type: string}
         rewardId: {type: string}
         name: {type: string}
         cost: {type: number}
         purchaseDate: {type: string}
    DiaryEntry:
       type: object
       properties:
         date: {type: string}
         initialReflection: {type: string}
         initialFeedback: {type: string}
         debrief: {type: string}
         finalFeedback: {type: string}
         grade: {type: string}
    Character:
       type: object
       properties:
         spent: {type: number}
         bonuses: {type: number}
...
    AtomicHabitsSuggestions:
      type: object
      properties:
        obvious: {type: array, items: {type: string}}
        attractive: {type: array, items: {type: string}}
        easy: {type: array, items: {type: string}}
        satisfying: {type: array, items: {type: string}}
    WeeklyGoalEvaluation:
      type: object
      properties:
        alignmentScore: {type: number}
        feedback: {type: string}
...
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content: 
          application/json:
            schema: 
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /goals:
    get:
      tags: [Goals]
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Goal'
    post:
      tags: [Goals]
      requestBody:
        content:
          application/json:
             schema:
               $ref: '#/components/schemas/Goal'
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
  /goals/{id}:
    put:
      tags: [Goals]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Goal'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
    delete:
      tags: [Goals]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        204: {description: Deleted}
  /weekly-goals:
    get:
      tags: [Goals]
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WeeklyGoal'
    post:
      tags: [Goals]
      requestBody:
        content:
          application/json:
            schema:
               $ref: '#/components/schemas/WeeklyGoal'
      responses:
        201:
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/WeeklyGoal'
  /weekly-goals/{id}:
    put:
      tags: [Goals]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      requestBody:
        content:
          application/json:
             schema:
               $ref: '#/components/schemas/WeeklyGoal'
      responses:
         200:
           content:
             application/json:
               schema:
                 $ref: '#/components/schemas/WeeklyGoal'
    delete:
      tags: [Goals]
      parameters:
        - in: path
          name: id
          required: true
          schema: {type: string}
      responses:
        204: {description: Deleted}
...
  /rewards:
     get:
       tags: [Rewards]
       responses:
         200:
           content:
             application/json:
               schema: {type: array, items: {$ref: '#/components/schemas/Reward'}}
     post:
       tags: [Rewards]
       requestBody:
         content: {application/json: {schema: {$ref: '#/components/schemas/Reward'}}}
       responses:
         201:
           content: {application/json: {schema: {$ref: '#/components/schemas/Reward'}}}
  /rewards/{id}:
    delete:
      tags: [Rewards]
      parameters: [{in: path, name: id, required: true, schema: {type: string}}]
      responses: {204: {description: Deleted}}
  /purchased-rewards:
     get:
       tags: [Rewards]
       responses:
         200:
           content:
             application/json:
               schema: {type: array, items: {$ref: '#/components/schemas/PurchasedReward'}}
     post:
       tags: [Rewards]
       requestBody:
         content: {application/json: {schema: {$ref: '#/components/schemas/PurchasedReward'}}}
       responses:
         201:
           content: {application/json: {schema: {$ref: '#/components/schemas/PurchasedReward'}}}
  /diary-entries:
     get:
       tags: [Diary]
       responses:
         200:
           content:
             application/json:
               schema: {type: array, items: {$ref: '#/components/schemas/DiaryEntry'}}
     post:
       tags: [Diary]
       requestBody:
         content: {application/json: {schema: {$ref: '#/components/schemas/DiaryEntry'}}}
       responses:
         201:
           content: {application/json: {schema: {$ref: '#/components/schemas/DiaryEntry'}}}
  /diary-entries/{date}:
      put:
        tags: [Diary]
        parameters: [{in: path, name: date, required: true, schema: {type: string}}]
        requestBody:
          content: {application/json: {schema: {$ref: '#/components/schemas/DiaryEntry'}}}
        responses:
          200:
            content: {application/json: {schema: {$ref: '#/components/schemas/DiaryEntry'}}}
  /character:
    get:
      tags: [User]
      responses:
         200:
           content:
             application/json:
               schema: {$ref: '#/components/schemas/Character'}
    put:
      tags: [User]
      requestBody:
         content: {application/json: {schema: {$ref: '#/components/schemas/Character'}}}
      responses:
        200:
           content: {application/json: {schema: {$ref: '#/components/schemas/Character'}}}
...